Justerer styling for del med bruker-modal
