+++ frontend/mulighetsrommet-veileder-flate/src/components/modal/Modal.module.scss
@@ -18,3 +18,11 @@
 .muted {
   color: var(--navds-semantic-color-text-muted);
 }
+
+.mt_0 {
+  margin-top: 0;
+}
+
+.mb_0 {
+  margin-bottom: 0;
+}
