+++ frontend/mulighetsrommet-veileder-flate/src/components/modal/delemodal/Delemodal.module.scss
@@ -1,8 +1,9 @@
 .delemodal {
-  background-color: red;
-
   &_tilbakemelding {
     text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
 
     .modal_btngroup {
       justify-content: space-around;
